Move revision module around and add license notices
