+++ pyshimmer/__init__.py
@@ -13,12 +13,14 @@
 
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 from .bluetooth.bt_api import ShimmerBluetooth
 from .bluetooth.bt_commands import DataPacket
 from .dev.base import DEFAULT_BAUDRATE
 from .dev.channels import ChannelDataType, EChannelType
 from .dev.exg import ExGMux, ExGRLDLead, ERLDRef, ExGRegister
 from .dev.fw_version import EFirmwareType
+from .dev.revisions import HardwareRevision, Shimmer3Revision
 from .reader.binary_reader import ShimmerBinaryReader
 from .reader.shimmer_reader import ShimmerReader
 from .uart.dock_api import ShimmerDock
